polar: Move away from inline imports
We want to avoid loading the entire SDK, because it doesn't get lazily loaded,
but having inline imports in every method does not look great either.

By splitting the client and implementation like this, we can get code
that is a bit nicer to read without getting a memory footprint in the
cases where we dont want it.
